Daniel Ankers <dan_a@manx.net>:
This patch against the stable version of envy24control gives access
to most of the Terratec DMX 6Fire mixer controls.  It also changes
the lables on the inputs in the monitor mixer to be more useful.
It should have no effect on your envy24control if you are using a
card other than the 6Fire.