envy24control wordclock text patch From: John Rigg <ad@sound-man.co.uk>
I reported a bug in envy24control about the text displayed when using
the BNC word clock input on a Delta 1010 (bug id 0001677). It says
`No signal' when locked and `Locked' when there's no signal. I wasn't
sure how to submit a patch using the Bugtrack interface, so I've
attached it here (this fixes it in alsa-tools-1.0.10).